[xDS] move XdsConfig out of XdsDependencyManager
