+++ config.w32
@@ -744,6 +744,7 @@ if (PHP_GRPC != "no") {
     "src\\core\\resolver\\resolver.cc " +
     "src\\core\\resolver\\resolver_registry.cc " +
     "src\\core\\resolver\\sockaddr\\sockaddr_resolver.cc " +
+    "src\\core\\resolver\\xds\\xds_config.cc " +
     "src\\core\\resolver\\xds\\xds_dependency_manager.cc " +
     "src\\core\\resolver\\xds\\xds_resolver.cc " +
     "src\\core\\server\\server.cc " +
