+++ gRPC-C++.podspec
@@ -1262,6 +1262,7 @@ Pod::Spec.new do |s|
                       'src/core/resolver/resolver_factory.h',
                       'src/core/resolver/resolver_registry.h',
                       'src/core/resolver/server_address.h',
+                      'src/core/resolver/xds/xds_config.h',
                       'src/core/resolver/xds/xds_dependency_manager.h',
                       'src/core/resolver/xds/xds_resolver_attributes.h',
                       'src/core/server/server.h',
@@ -2548,6 +2549,7 @@ Pod::Spec.new do |s|
                               'src/core/resolver/resolver_factory.h',
                               'src/core/resolver/resolver_registry.h',
                               'src/core/resolver/server_address.h',
+                              'src/core/resolver/xds/xds_config.h',
                               'src/core/resolver/xds/xds_dependency_manager.h',
                               'src/core/resolver/xds/xds_resolver_attributes.h',
                               'src/core/server/server.h',
