+++ gRPC-Core.podspec
@@ -1929,6 +1929,8 @@ Pod::Spec.new do |s|
                       'src/core/resolver/resolver_registry.h',
                       'src/core/resolver/server_address.h',
                       'src/core/resolver/sockaddr/sockaddr_resolver.cc',
+                      'src/core/resolver/xds/xds_config.cc',
+                      'src/core/resolver/xds/xds_config.h',
                       'src/core/resolver/xds/xds_dependency_manager.cc',
                       'src/core/resolver/xds/xds_dependency_manager.h',
                       'src/core/resolver/xds/xds_resolver.cc',
@@ -3327,6 +3329,7 @@ Pod::Spec.new do |s|
                               'src/core/resolver/resolver_factory.h',
                               'src/core/resolver/resolver_registry.h',
                               'src/core/resolver/server_address.h',
+                              'src/core/resolver/xds/xds_config.h',
                               'src/core/resolver/xds/xds_dependency_manager.h',
                               'src/core/resolver/xds/xds_resolver_attributes.h',
                               'src/core/server/server.h',
