+++ src/core/BUILD
@@ -5858,6 +5858,7 @@ grpc_cc_library(
         "unique_type_name",
         "xds_cluster",
         "xds_common_types",
+        "xds_config",
         "xds_dependency_manager",
         "xds_health_status",
         "//:config",
@@ -5923,8 +5924,8 @@ grpc_cc_library(
         "resolved_address",
         "subchannel_interface",
         "validation_errors",
+        "xds_config",
         "xds_credentials",
-        "xds_dependency_manager",
         "xds_endpoint",
         "//:call_tracer",
         "//:config",
@@ -6636,7 +6637,7 @@ grpc_cc_library(
         "resolved_address",
         "subchannel_interface",
         "validation_errors",
-        "xds_dependency_manager",
+        "xds_config",
         "xds_health_status",
         "//:config",
         "//:debug_location",
@@ -6982,6 +6983,31 @@ grpc_cc_library(
 )
 
 grpc_cc_library(
+    name = "xds_config",
+    srcs = [
+        "resolver/xds/xds_config.cc",
+    ],
+    hdrs = [
+        "resolver/xds/xds_config.h",
+    ],
+    external_deps = [
+        "absl/container:flat_hash_map",
+        "absl/status:statusor",
+        "absl/strings",
+        "absl/types:variant",
+    ],
+    language = "c++",
+    deps = [
+        "match",
+        "ref_counted",
+        "xds_cluster",
+        "xds_endpoint",
+        "xds_listener",
+        "xds_route_config",
+    ],
+)
+
+grpc_cc_library(
     name = "xds_dependency_manager",
     srcs = [
         "resolver/xds/xds_dependency_manager.cc",
@@ -7006,6 +7032,7 @@ grpc_cc_library(
         "xds_endpoint",
         "xds_listener",
         "xds_route_config",
+        "xds_config",
         "//:config",
         "//:gpr",
         "//:grpc_resolver",
@@ -7057,6 +7084,7 @@ grpc_cc_library(
         "xds_http_filter",
         "xds_listener",
         "xds_route_config",
+        "xds_config",
         "xxhash_inline",
         "//:channel_arg_names",
         "//:config",
