+++ src/python/grpcio/grpc_core_dependencies.py
@@ -753,6 +753,7 @@ CORE_SOURCE_FILES = [
     'src/core/resolver/resolver.cc',
     'src/core/resolver/resolver_registry.cc',
     'src/core/resolver/sockaddr/sockaddr_resolver.cc',
+    'src/core/resolver/xds/xds_config.cc',
     'src/core/resolver/xds/xds_dependency_manager.cc',
     'src/core/resolver/xds/xds_resolver.cc',
     'src/core/server/server.cc',
